Add login and register api

+++ src/main/java/com/picfood/server/entity/User.java
@@ -13,7 +13,7 @@ import java.util.Date;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer userId;
+    private Long userId;
 
     private String name;
     private String avatar;
@@ -30,11 +30,11 @@ public class User {
 
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
